% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.survstan}
\alias{anova.survstan}
\title{anova method for survstan models}
\usage{
\method{anova}{survstan}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\value{
the ANOVA table.
}
\description{
Compute analysis of variance (or deviance) tables for one or more fitted model objects.
}
