% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein.R
\name{bernstein}
\alias{bernstein}
\title{Bernstein polynomial}
\usage{
bernstein(m = NULL)
}
\arguments{
\item{m}{the Bernstein polynomial's degree; default is NULL.}
}
\value{
a list with the baseline name and the polynomial's degree m.
}
\description{
This function is used to allow the user to specify an arbitrary value for the polynomial's degree m. If m = NULL, then m = min(m_max, ceiling(n^0.4)) is used, where m_max = 15.
}
