% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_time.R
\name{cross_time}
\alias{cross_time}
\title{Generic S3 method cross_time}
\usage{
cross_time(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the crossing survival time
}
\description{
Generic S3 method cross_time
}
