% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{ggresiduals}
\alias{ggresiduals}
\title{Generic S3 method ggresiduals}
\usage{
ggresiduals(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the desired residual plot.
}
\description{
Generic S3 method ggresiduals
}
\details{
Generic method to plot residuals of survival models.
}
