% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{model.matrix.survstan}
\alias{model.matrix.survstan}
\title{Model.matrix method for survstan models}
\usage{
\method{model.matrix}{survstan}(object, ...)
}
\arguments{
\item{object}{an object of the class survstan.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The model matrix (or matrices) for the fit.
}
\description{
Reconstruct the model matrix for a survstan model.
}
