% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\title{Piecewise exponential baseline}
\usage{
piecewise(rho = NULL, m = NULL)
}
\arguments{
\item{rho}{the specified time grid; default is NULL.}

\item{m}{the number of intervals; default is NULL.}
}
\value{
a list with the baseline name, and the provided time grid and number of intervals.
}
\description{
This function is used to allow the user to specify the piecewise exponential baseline with arbitrary time grid/number of intervals.
}
