% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.survstan}
\alias{print.summary.survstan}
\title{Print the summary.survstan output}
\usage{
\method{print}{summary.survstan}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.survstan.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects.
}
\description{
Produces a printed summary of a fitted survstan model.
}
