% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{se}
\alias{se}
\title{Generic S3 method se}
\usage{
se(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the standard errors associated with a set of parameter estimators for a given model.
}
\description{
Generic S3 method se
}
