% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{se.survstan}
\alias{se.survstan}
\title{Estimated standard errors}
\usage{
\method{se}{survstan}(object, ...)
}
\arguments{
\item{object}{an object of the class survstan.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector with the standard errors.
}
\description{
Estimated standard errors
}
\examples{
\donttest{
library(survstan)
fit <- aftreg(Surv(futime, fustat) ~ ecog.ps + rx, data = ovarian, baseline = "weibull", init = 0)
se(fit)
}

}
