% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.survstan}
\alias{summary.survstan}
\title{Summary for a survstan object}
\usage{
\method{summary}{survstan}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{an object of the class 'survstan'.}

\item{conf.level}{the confidence level required.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for a survstan object
}
