% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{Generic S3 method tidy}
\usage{
tidy(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{conf.level}{the confidence level required.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a tibble with a summary of the fit.
}
\description{
Generic S3 method tidy
}
\details{
Convert a fitted model into a tibble.
}
