% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{coef.survstan}
\alias{coef.survstan}
\title{Estimated regression coefficients}
\usage{
\method{coef}{survstan}(object, ...)
}
\arguments{
\item{object}{an object of the class survstan}

\item{...}{further arguments passed to or from other methods}
}
\value{
the estimated regression coefficients
}
\description{
Estimated regression coefficients
}
\examples{
\donttest{
library(survstan)
fit <- aftreg(Surv(futime, fustat) ~ ecog.ps + rx, data = ovarian, baseline = "weibull", init = 0)
coef(fit)
}

}
