% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{emmeans-survstan-helpers}
\alias{emmeans-survstan-helpers}
\alias{recover_data.survstan}
\alias{recover_data.ypreg}
\title{Support Functions for \pkg{emmeans}}
\usage{
recover_data.survstan(object, ...)

recover_data.ypreg(object, term = c("short", "long"), ...)
}
\arguments{
\item{object}{An object of the same class as is supported by a new method.}

\item{...}{Additional parameters that may be supported by the method.}

\item{term}{character specifying whether short or long term regression coefficients are to be used.}
}
\description{
Functions required for compatibility of \pkg{survstan} with \pkg{emmeans}.
Users are not required to call these functions themselves. Instead,
they will be called automatically by the \code{emmeans} function
of the \pkg{emmeans} package.
}
