% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.survstan}
\alias{summary.survstan}
\title{Summary for a survstan object}
\usage{
\method{summary}{survstan}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{the result of a call to summary.survstan}

\item{conf.level}{the confidence level required.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an object of the class summary.survstan containing a summary of the fitted model.
}
\description{
Summary for a survstan object
}
