% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survstan-package.R
\name{survstan-package}
\alias{survstan-package}
\alias{survstan}
\title{The 'survstan' package.}
\description{
The aim of the R package survstan is to provide a toolkit for fitting survival models using Stan. The R package survstan can be used to fit right-censored survival data under independent censoring. The implemented models allow the fitting of survival data in the presence/absence of covariates. All inferential procedures are currently based on the maximum likelihood (ML) approach.

_PACKAGE
}
\references{
\insertRef{rstan}{survstan}

\insertRef{bookLawless}{survstan}

\insertRef{1983Bennett}{survstan}

\insertRef{2000Chen}{survstan}

\insertRef{2021_Demarqui}{survstan}
}
